# Switch from Puppeteer to Playwright for reliable and concurrent browser tests