--- a/test/helpers/with-page.js
+++ b/test/helpers/with-page.js
@@ -1,7 +1,7 @@
-import puppeteer from 'puppeteer';
+import {chromium} from 'playwright-chromium';
 
 export default async function withPage(t, run) {
-	const browser = await puppeteer.launch();
+	const browser = await chromium.launch();
 	const page = await browser.newPage();
 	try {
 		await run(t, page);