@sindresorhus 2021-01-16
Retry tests seem to be failing on Node.js 14 though.