@sholladay 2021-01-16
The tests in `test/retry.js` don't use Puppeteer and definitely shouldn't be affected by these changes. I think GitHub Actions is just having a bad day. I re-ran the job and it passed. I haven't had a single issue locally and I've run the tests many times.