@sindresorhus 2021-01-16
This is great! I didn't realize Playwright is so much faster.